+++ prisma/migrations/20260211151842_scores_duration_player_not_null/migration.sql
@@ -1,0 +1,10 @@
+/*
+  Warnings:
+
+  - Made the column `player_id` on table `scores` required. This step will fail if there are existing NULL values in that column.
+  - Made the column `duration` on table `scores` required. This step will fail if there are existing NULL values in that column.
+
+*/
+-- AlterTable
+ALTER TABLE "scores" ALTER COLUMN "player_id" SET NOT NULL,
+ALTER COLUMN "duration" SET NOT NULL;