+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.3.4b -
 
 - Update Vercel CLI to 50.15.1 and AWS-SDK
+- Update DB schema: `duration` and `player_id` converted to `NOT NULL`
+- `POST /points`: Update `duration` to required and positive number
 
 ## 0.3.3b - 10/02/2026 19:10 UTC+1